fix(core,lineareq,saturation,multiband,oversampling,limiter): downstream warning hygiene + a header-hygiene gate
TabbyEQ links its test targets against juce_recommended_warning_flags and
our headers warned there while core's own build was silent. Fixed at the
root and gated so the class of bug dies in core CI, not downstream:

- lineareq/NaturalPhaseEq computeTaper(): the taper mixed float/double
  (kPi is double) and narrowed silently into the float buffer. Now computed
  honestly in double with ONE explicit (float) cast — the house form used
  three lines above (-Wimplicit-float-conversion).
- core/Math.h: felitronics::core::exactlyEqual<T> — a JUCE-free analogue of
  juce::exactlyEqual (pragma-scoped -Wfloat-equal suppression around one
  intentional exact ==). LinearSmoother::approxEqual's non-finite branch now
  uses it; the juce::approximatelyEqual-parity semantics are unchanged.
- Explicit-cast cleanups the strict set surfaced: (size_t)a * int products
  in MultibandProcessor / PolyphaseOversampler / TruePeakLimiter
  (-Wsign-conversion) and double->float narrows in Saturator::applyParams
  (-Wimplicit-float-conversion).

The gate: test_support/HeaderHygiene.cpp — one TU including EVERY public
header (felitronics + the teq shims TabbyEQ actually includes), key class
templates explicitly instantiated so warnings inside template bodies fire,
compiled as target felitronics_header_hygiene under -Werror with a
downstream-grade set (-Wall -Wextra -Wpedantic -Wconversion
-Wsign-conversion -Wfloat-equal -Wshadow -Wcast-align -Wswitch-enum
-Wzero-as-null-pointer-constant + clang extras -Wshorten-64-to-32
-Wimplicit-float-conversion ...). Chosen over strict flags on the test
targets: the tests' exact float == checks (bit-exact/reference-NULL
discipline) are intentional; the hygiene TU checks exactly what a consumer
compiles — the headers.

Verified: hygiene target compiles with 0 warnings (clang, -Werror); full
ctest 29 targets green.

+++ modules/saturation/include/felitronics/saturation/Saturator.h
@@ -151,10 +151,11 @@ private:
         const float dcHz     = finite (params_.dcBlockHz, 10.0f);
         shaper_.setShape (params_.shape);
         shaper_.setBias  (bias);
-        shaper_.setDrive (core::dbToGain (driveDb) - 1.0f);             // driveDb 0 → k≈0 (linear)
-        comp_    = std::pow (std::max (1.0e-6f, shaper_.slopeAtZero()), -std::clamp (autoComp, 0.0f, 1.0f));
+        shaper_.setDrive ((float) (core::dbToGain (driveDb) - 1.0));    // driveDb 0 → k≈0 (linear)
+        comp_    = (float) std::pow ((double) std::max (1.0e-6f, shaper_.slopeAtZero()),
+                                     (double) -std::clamp (autoComp, 0.0f, 1.0f));
         mix_     = std::clamp (finite (params_.mix, 1.0f), 0.0f, 1.0f);
-        outGain_ = core::dbToGain (finite (params_.outputDb, 0.0f));
+        outGain_ = (float) core::dbToGain (finite (params_.outputDb, 0.0f));
         const double fsOs = fs_ * (double) os_;
         const double fc   = std::clamp ((double) dcHz, 0.0, 0.49 * fsOs);
         dcR_ = (fc <= 0.0) ? 0.0f : (float) std::exp (-2.0 * core::kPi * fc / fsOs);
